lib multi

+++ lib/multi.c
@@ -343,7 +343,7 @@ static size_t hash_fd(void *key, size_t key_length, size_t slots_num)
   curl_socket_t fd = *((curl_socket_t *) key);
   (void) key_length;
 
-  return (fd % (curl_socket_t)slots_num);
+  return (size_t)(fd % (curl_socket_t)slots_num);
 }
 
 /*
